Move test-only utils from core to tests/test_utils.py
- Move format_task_data, get_system_uuid, get_related_object_safely
  to tests/test_utils.py since they're only used in tests
- Keep build_error_response, get_count_safely in apps/core/utils.py
  as they're used by production API code
- Fix stale imports referencing old metrics_service command location
