+++ tests/unit/core/test_init_service_id_command.py
@@ -7,7 +7,7 @@ from io import StringIO
 import pytest
 from django.test import TestCase
 
-from apps.core.management.commands.metrics_service import Command
+from apps.tasks.management.commands.metrics_service import Command
 
 
 @pytest.mark.django_db
