+++ tests/unit/core/test_metrics_service_command.py
@@ -11,7 +11,7 @@ from unittest.mock import MagicMock, call, patch
 import pytest
 from django.test import TestCase
 
-from apps.core.management.commands.metrics_service import Command
+from apps.tasks.management.commands.metrics_service import Command
 
 
 class TestMetricsServiceCommand(TestCase):
